refactor: remove app specific components from interpretations folder
